Moved BankTest and PlayerTest and minor changes
