+++ src/test/java/player/BankTest.java
@@ -1,5 +1,5 @@
 
-package game;
+package player;
 import org.junit.Assert;
 import org.junit.Test;
